recommender-back/pyproject.toml
[tool.poetry]
name = "recommender-back"
version = "0.1.0"
description = ""
authors = ["ruusukivi <ruusukivi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.3.2"
fmiopendata = {git = "https://github.com/pnuu/fmiopendata.git"}
flask-cors = "^3.0.10"
invoke = "^2.1.2"
overpy = "^0.6"
datetime = "^5.1"
suntime = "^1.2.5"
flask-caching = "^2.0.2"
redis = "^3.5.3"
e = "^1.4.5"
eccodes = "^1.5.2"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
gunicorn = "^20.1"
coverage = "^7.2.7"
celery = "^5.3.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
autopep8 = "^2.0.2"
pylint = "^2.17.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
